Updated experimental VisionOS version with the changes made during these years to the core engine, so the app compiles and runs in a simulator.

+++ SlipNFrag-VisionOS/LightmapCopying.h
@@ -11,7 +11,7 @@
 struct LightmapCopying
 {
 	uint32_t lightmap;
-	uint16_t* data;
+	uint32_t* data;
 	uint32_t width;
 	uint32_t height;
 	uint32_t size;
